Split compound.* into multiple files by domain
By relocating according classes into compound_registry.*, compound_absorber_system.* and compound_emitter_system.*, issue #116 should be resolved.

- CompoundMovementSystem was kept in compound.*, because it is directly and only(*) linked to the CompoundComponent.
(*in relation to everything else that was located in compound.*)
