+++ src/microbe_stage/CMakeLists.txt
@@ -1,6 +1,12 @@
 add_sources(
     ${CMAKE_CURRENT_SOURCE_DIR}/compound.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/compound.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/compound_registry.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/compound_registry.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/compound_absorber_system.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/compound_absorber_system.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/compound_emitter_system.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/compound_emitter_system.h
     ${CMAKE_CURRENT_SOURCE_DIR}/bio_process_registry.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/bio_process_registry.h
     ${CMAKE_CURRENT_SOURCE_DIR}/script_bindings.cpp
