+++ src/microbe_stage/bio_process_registry.cpp
@@ -1,6 +1,7 @@
 
 #include "microbe_stage/bio_process_registry.h"
 #include "microbe_stage/compound.h"
+#include "microbe_stage/compound_registry.h"
 #include "scripting/luabind.h"
 
 #include "tinyxml.h"
